cleanup & move assumptions to another header
